Prepare buildpack for automatic dependency updates
- In order to add the dotnet dependency to the pipeline which
automatically watches for new versions, builds them, and adds them to
the buildpack, we need to make some changes so that only YAML files are
updated, not Ruby code

- Add default_versions section to manifest.yml and use the compile
extensions script to select the default version

- Move the mapping of .NET SDK -> framework versions into a new file,
'dotnet-versions.yml'. Use this file to select the proper SDK version
given a *.runtimeconfig.json file

- Update the constructor of DotnetVersion so that the manifest.yml and
dotnet-versions.yml files are passed in. This allows the specs to define
their own contents for these files

[#130932637]

Signed-off-by: Anna Thornton <athornton@pivotal.io>